requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
baidu-aip==4.16.13
certifi==2026.2.25
chardet==7.4.1
charset-normalizer==3.4.7
distro==1.9.0
dotenv==0.9.9
edge-tts==7.2.8
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
multidict==6.7.1
openai==2.30.0
packaging==26.0
propcache==0.4.1
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
python-dotenv==1.2.2
requests==2.33.1
setuptools==82.0.1
sniffio==1.3.1
tabulate==0.10.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
yarl==1.23.0
